Let ActivatePeer do less work
And pass `ConstructorInfo` instead of types information.
