+++ src/Java.Interop/Java.Interop/JniRuntime.JniValueManager.cs
@@ -86,7 +86,7 @@ namespace Java.Interop
 
 			public abstract List<JniSurfacedPeerInfo>   GetSurfacedPeers ();
 
-			public abstract void ActivatePeer (IJavaPeerable? self, JniObjectReference reference, Type peerType, Type []? constructorArguments, object? []? argumentValues);
+			public abstract void ActivatePeer (IJavaPeerable? self, JniObjectReference reference, ConstructorInfo cinfo, object? []? argumentValues);
 
 			public void ConstructPeer (IJavaPeerable peer, ref JniObjectReference reference, JniObjectReferenceOptions options)
 			{
