+++ tests/Java.Interop-Tests/Java.Interop/JniRuntime.JniValueManagerTests.cs
@@ -1,4 +1,5 @@
-﻿using System;
+using System;
+using System.Reflection;
 using System.Collections.Generic;
 
 using Java.Interop;
@@ -60,7 +61,7 @@ namespace Java.InteropTests {
 				return null;
 			}
 
-			public override void ActivatePeer (IJavaPeerable self, JniObjectReference reference, Type peerType, Type [] constructorArguments, object [] argumentValues)
+			public override void ActivatePeer (IJavaPeerable self, JniObjectReference reference, ConstructorInfo cinfo, object [] argumentValues)
 			{
 				throw new NotImplementedException ();
 			}
